clean up Leftbar component
