--- spa/src/components/account/common/leftbar/Content.ts
@@ -1,25 +1,0 @@
-import { ICONS } from 'assets/icons/SvgIcon';
-
-const Content = {
-  heading: 'We’ve helped raise millions for our clients',
-  advantages: [
-    {
-      svg: ICONS.ACCOUNT_ACC_BALANCE,
-      heading: 'Save time, money, and democracy',
-      subheading: 'Sustain your newsroom through <i>voluntary</i> contributions'
-    },
-    {
-      svg: ICONS.ACCOUNT_NEWSROOM,
-      heading: 'For newsrooms, by newsrooms',
-      subheading: 'An <i>affordable</i>, DIY contributions management software'
-    },
-    { svg: ICONS.ACCOUNT_CART, heading: 'No CRM required', subheading: 'Create a checkout page in <i>minutes</i>' },
-    {
-      svg: ICONS.ACCOUNT_UNFOLD_MORE,
-      heading: 'Strategic integrations',
-      subheading: '<i>Real-time</i> payment notifications & payment processing'
-    }
-  ]
-};
-
-export default Content;
